# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Voyager CRM",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://voyager.crm.localhost"
        }
    ],
    "paths": {
        "/api/leads/": {
            "get": {
                "tags": [
                    "Lead"
                ],
                "summary": "Show Leads",
                "description": "Returns list of all Leads",
                "responses": {
                    "200": {
                        "description": "Successful operation."
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "500": {
                        "description": "CRM failure."
                    }
                }
            }
        },
        "/api/leads": {
            "post": {
                "tags": [
                    "Lead"
                ],
                "summary": "Store",
                "description": "Store Lead",
                "requestBody": {
                    "description": "Lead information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Lead"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lead"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "500": {
                        "description": "CRM failure."
                    }
                }
            }
        },
        "/api/leads/{id}": {
            "get": {
                "tags": [
                    "Lead"
                ],
                "summary": "Show",
                "description": "Returns Lead by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Put existing Lead ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 12
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lead"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "500": {
                        "description": "CRM failure."
                    }
                }
            },
            "put": {
                "tags": [
                    "Lead"
                ],
                "summary": "Update",
                "description": "Update Lead",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Put existing Lead ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 12
                        }
                    }
                ],
                "requestBody": {
                    "description": "Lead update information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Lead"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lead"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "500": {
                        "description": "CRM failure."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Lead"
                ],
                "summary": "Delete",
                "description": "Delete Lead",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Put existing Lead ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 12
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation."
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "500": {
                        "description": "CRM failure."
                    }
                }
            }
        },
        "/api/steps/": {
            "get": {
                "tags": [
                    "Step"
                ],
                "summary": "Show Steps",
                "description": "Returns list of all Steps",
                "responses": {
                    "200": {
                        "description": "Mostrar Steps"
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "500": {
                        "description": "CRM failure."
                    }
                }
            }
        },
        "/api/steps": {
            "post": {
                "tags": [
                    "Step"
                ],
                "summary": "Store",
                "description": "Store Step",
                "requestBody": {
                    "description": "Step information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Step"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Step"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "500": {
                        "description": "CRM failure."
                    }
                }
            }
        },
        "/api/steps/{id}": {
            "get": {
                "tags": [
                    "Step"
                ],
                "summary": "Show",
                "description": "Returns Step by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Put existing step ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 7
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Step"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "500": {
                        "description": "CRM failure."
                    }
                }
            },
            "put": {
                "tags": [
                    "Step"
                ],
                "summary": "Update",
                "description": "Update Step",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Put existing Step ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 7
                        }
                    }
                ],
                "requestBody": {
                    "description": "Store Step body",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Step"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Step"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "500": {
                        "description": "CRM failure."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Step"
                ],
                "summary": "Delete",
                "description": "Delete Step",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Put existing Step ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 7
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation."
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "500": {
                        "description": "CRM failure."
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Lead": {
                "title": "Lead",
                "description": "Lead model",
                "properties": {
                    "name": {
                        "title": "Lead Name",
                        "description": "Lead Name",
                        "type": "string",
                        "example": "Pedro"
                    },
                    "email": {
                        "title": "Lead email",
                        "description": "Lead email",
                        "type": "string",
                        "example": "pedro@ejemplo1.com"
                    },
                    "phone": {
                        "title": " Lead phone",
                        "description": "Lead phone",
                        "type": "string",
                        "example": "1233567"
                    },
                    "message": {
                        "title": "Message for Lead",
                        "description": "Message for Lead",
                        "type": "string",
                        "example": "Hello Pedro"
                    },
                    "step_id": {
                        "title": "Step ID for Lead",
                        "description": "Step ID for Lead",
                        "type": "int",
                        "example": 3
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Lead"
                }
            },
            "Step": {
                "title": "Step",
                "description": "Step model",
                "properties": {
                    "key": {
                        "title": "Step Key",
                        "description": "Step Key",
                        "type": "string",
                        "example": "abierto"
                    },
                    "value": {
                        "title": "Step Value",
                        "description": "Step Value",
                        "type": "string",
                        "example": "Abierto"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Step"
                }
            }
        }
    },
    "tags": [
        {
            "name": "Lead",
            "description": "Lead Operations"
        },
        {
            "name": "Step",
            "description": "Step Operations"
        }
    ]
}